@@ -1,4 +1,4 @@
-name: ticketing_2
+name: ticketing_app
 description: A new Flutter application.
 
 # The following defines the version and build number for your application.
@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   datetime_picker_formfield: ^1.0.0
