@@ -4,16 +4,26 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'dart:io';
 
-class TicketFormPage extends StatefulWidget {
-  TicketFormPage({this.title});
-  final String title;
+import 'package:ticketing_app/util/constants.dart';
+
+class CreateNewTicketScreen extends StatefulWidget {
+//  TicketFormPage({this.title,this.canEdit});
+//  final String title;
+//  final bool canEdit;
 
   @override
-  _TicketFormPageState createState() => _TicketFormPageState();
+  _CreateNewTicketScreenState createState() => _CreateNewTicketScreenState();
 }
 
-class _TicketFormPageState extends State<TicketFormPage> {
+class _CreateNewTicketScreenState extends State<CreateNewTicketScreen> {
   File _image;
+  String probDesc;
+  String partNo;
+  String serialNo;
+  DateTime date;
+  String contactNum;
+
+
 
   Future getImage() async {
     var image = await ImagePicker.pickImage(source: ImageSource.gallery);
@@ -22,13 +32,15 @@ class _TicketFormPageState extends State<TicketFormPage> {
     });
   }
 
-
   final GlobalKey<FormState> _formKey = new GlobalKey<FormState>();
   final format = DateFormat("yyyy-MM-dd");
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Create New Ticket'),),
+      appBar: AppBar(
+        title: Text('Create New Ticket'),
+      ),
       body: new SafeArea(
         child: Form(
           key: _formKey,
@@ -41,6 +53,15 @@ class _TicketFormPageState extends State<TicketFormPage> {
                   labelText: 'Problem Description',
                   icon: Icon(Icons.person),
                 ),
+                validator: (value) {
+                  if (value.isEmpty) {
+                    return 'Please add some description to your issue';
+                  }
+                  return null;
+                },
+                onSaved: (value){
+                  probDesc = value;
+                },
               ),
               TextFormField(
                 decoration: InputDecoration(
@@ -48,7 +69,6 @@ class _TicketFormPageState extends State<TicketFormPage> {
                   icon: Icon(Icons.subdirectory_arrow_right),
                 ),
               ),
-
               TextFormField(
                 decoration: InputDecoration(
                   labelText: 'Serial No',
@@ -64,8 +84,8 @@ class _TicketFormPageState extends State<TicketFormPage> {
                       firstDate: DateTime(1900),
                       initialDate: currentValue ?? DateTime.now(),
                       lastDate: DateTime(2100));
-                },),
-
+                },
+              ),
               TextFormField(
                 decoration: InputDecoration(
                   labelText: 'Contact Details',
@@ -81,17 +101,32 @@ class _TicketFormPageState extends State<TicketFormPage> {
                 ),
               ),
               SizedBox(
-                  width: 20.0,
-                  height: 20.0,
-                  child: RaisedButton(
-                    onPressed: getImage,
-                    child: Icon(Icons.add_a_photo),
-                  )),
+                width: 20.0,
+                height: 20.0,
+                child: RaisedButton(
+                  onPressed: getImage,
+                  child: Icon(Icons.add_a_photo),
+                ),
+              ),
+              RaisedButton(
+                color: kSubmitButton,
+                textColor: kTextButton,
+                onPressed: () {
+                  // Validate returns true if the form is valid, otherwise false.
+                  if (_formKey.currentState.validate()) {
+                    // If the form is valid, display a snackbar. In the real world,
+                    // you'd often call a server or save the information in a database.
+
+                   print('Rest api call');
+                   print(probDesc);
+                  }
+                },
+                child: Text('Submit'),
+              ),
             ],
           ),
-
-        ),),
+        ),
+      ),
     );
   }
-
 }
